% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{check_key}
\alias{check_key}
\title{Check whether API key is valid}
\usage{
check_key(key = NULL)
}
\arguments{
\item{key}{(Optional) A string. Your NewsAPI key.}
}
\description{
Checks whether an API key is valid. Prints a message to let the user know whether
their key is usable with NewsAPI (case in which the function prints "All OK"),
or whether an error occurred (case in which the function prints a descriptive
message of said error).
}
