% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couleurscharte.R
\docType{data}
\name{couleurscharte}
\alias{couleurscharte}
\title{Les couleurs de la charte du Parisien}
\format{
Un tableur aec 45 lignes et 3 colonnes
}
\usage{
data(couleurscharte)
}
\description{
Les couleurs associées aux design system du Parisien
}
\details{
\itemize{
\item code. le css dans le DS
\item hexa. le code hexadécimal associé
\item couleursimple. Les couleurs en français (bleu, vert, ou NA si pas utile)
}
}
\keyword{datasets}
