% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_part.R
\name{extract_part}
\alias{extract_part}
\title{extract_part Title}
\usage{
extract_part(x, sep, partie)
}
\arguments{
\item{x}{Une chaine de caractères}

\item{sep}{Une chaine de caractères séparant deux parties}

\item{partie}{Un chiffre indiquant quelle partie on veut récupérer}
}
\value{
Une chaine de caractères
}
\description{
extract_part Title
}
\examples{
Coucou<-"Amendement n°3"
extract_part(x=Coucou,sep="°",partie=2)
}
