% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENDPOINT-sources.R
\name{get_sources}
\alias{get_sources}
\title{Get index of available sources}
\usage{
get_sources(category = NULL, language = NULL, country_iso = NULL)
}
\arguments{
\item{category}{(Optional) A string. Find sources that display news of a specific category,
such as "business", "entertainment", or "sports".}

\item{language}{(Optional) A string. Find sources that display news in a specific language.
Possible options are: Arabic, German, English, Spanish, French, Hebrew, Italian
Dutch, Norwegian, Portuguese, Russian, Swedish, Urdu, and Chinese.
By default, the API returns sources in all languages.}

\item{country_iso}{(Optional) A string. The 2-letter ISO 3166-1 code of the country you
want to get headlines for, such as "US" or "FR". This parameter cannot be mixed with
the sources parameter.}
}
\value{
A list of news publishers.
}
\description{
Returns the subset of news publishers that top headlines are available from.
It's mainly a convenience function that can be used to keep track of the
publishers available on the API.
}
\examples{
get_sources(language = "de")

}
