% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top-headlines.R
\name{get_top_headlines}
\alias{get_top_headlines}
\title{Get top headlines articles}
\usage{
get_top_headlines(
  keyword = NULL,
  country = NULL,
  category = NULL,
  sources = NULL,
  page_size = NULL,
  page = NULL
)
}
\arguments{
\item{keyword}{(Optional) A string. Keyword(s) based on which to search for articles,
such as "Hong Kong protests" or "Amazon forest".}

\item{country}{(Optional) A string. The 2-letter ISO 3166-1 code of the country you
want to get headlines for, such as "US" or "FR". This parameter cannot be mixed with
the sources parameter.}

\item{category}{(Optional) A string. The category you want to get top headlines for, such
as "business", "entertainment", or "sports". This parameter cannot be mixed with
the sources parameter.}

\item{sources}{(Optional) A character vector. The news sources or blogs you want top
headlines from. See the get_sources() function for how to obtain a full list
of the available sources. This parameter cannot be mixed with the country_iso
or category parameters.}

\item{page_size}{(Optional) An integer. The number of results to return per request.
The API's default is 20, and the maximum is 100.}

\item{page}{(Optional) An integer. Use this to page through the results if the total
results found is greater than the page size.}
}
\value{
A data frame of article information, with source, author, title,
description, URL, publication date, and a short preview of the content.
}
\description{
Returns a list of current top headlines articles based on some query. At least
one parameter from among keyword, country_iso, category, and sources must be
specified, with the limitation that the sources parameter cannot be mixed with
country_iso or category.
}
\examples{
get_top_headlines(keyword = "Panama Canal", page_size = 10, page = 2)
get_top_headlines(country = "br", category = "sports")
get_top_headlines(keyword = "iPhone 10", source = "techcrunch")

}
