% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%!in\%}
\alias{\%!in\%}
\title{notin Title}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{Le vecteur initial}

\item{y}{Le vecteur dans lequel vérifier l'absence de x}
}
\value{
un vecteur avec les indices des x absents de y
}
\description{
notin Title
}
\examples{
1:10 \%!in\% 20:6
}
