% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{has_key}
\alias{has_key}
\title{Check whether an API key has been registered}
\usage{
has_key()
}
\value{
A boolean indicating whether an API key has been registered within the
current R session.
}
\description{
Checks whether an API key has been registered within the current R session,
either manually by using register_key(), or automatically from the .Renviron
file.
}
