% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majuscule.R
\name{majuscule}
\alias{majuscule}
\title{majuscule Title}
\usage{
majuscule(x)
}
\arguments{
\item{x}{Une chaine de caractères}
}
\value{
Une chaine de caractèreq avec une majuscule
}
\description{
majuscule Title
}
\examples{
majuscule("mon petit chat")
majuscule("ATTENTION DERRIERE TOI !")
}
