% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{notin}
\alias{notin}
\title{notin notin}
\usage{
notin(x, y)
}
\arguments{
\item{x}{Le vecteur initial}

\item{y}{Le vecteur dans lequel vérifier l'absence de x}
}
\value{
un vecteur avec les indices des x absents de y
}
\description{
notin notin
}
\examples{
notin(1:10,20:6)
}
