% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuancier.R
\name{nuancier}
\alias{nuancier}
\title{nuancier Title}
\usage{
nuancier(couldep, coularr, nombre)
}
\arguments{
\item{couldep}{Couleur de départ du dégradé}

\item{coularr}{Couleur d'arrivée du dégradé}

\item{nombre}{Nombre de couleurs entre les deux}
}
\value{
un vecteur contenant une liste de couleurs
}
\description{
nuancier Title
}
\examples{
nuancier("white","#CC2828",5)
}
