% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortunjpeg.R
\name{sortunjpeg}
\alias{sortunjpeg}
\title{sortunjpeg sortunjpeg}
\usage{
sortunjpeg(graph, largeur, hauteur, titre)
}
\arguments{
\item{graph}{L'objet à afficher dans le jpeg}

\item{largeur}{La largeur en pixels}

\item{hauteur}{La hauteur en pixels}

\item{titre}{Le nom du jpeg qui sera exporté, sans l'extension}
}
\value{
un fichier contenant l'objet appelé image
}
\description{
sortunjpeg sortunjpeg
}
\examples{
library(dplyr)
library(ggplot2)
sampledata<-data.frame(x=1:10,y=1:10)
graphique<-ggplot(sampledata)+
  geom_point(aes(x,y))
\dontrun{
sortunjpeg(graphique,200,200,"graphiquedetest")
}

}
