% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortunsvg.R
\name{sortunsvg}
\alias{sortunsvg}
\title{sortunsvg sortunsvg}
\usage{
sortunsvg(graph, largeur, hauteur, titre)
}
\arguments{
\item{graph}{L'objet à afficher dans le svg}

\item{largeur}{La largeur en points}

\item{hauteur}{La hauteur en points}

\item{titre}{Le nom du svg qui sera exporté, sans l'extension}
}
\value{
un fichier contenant l'objet appelé image
}
\description{
sortunsvg sortunsvg
}
\examples{
library(dplyr)
library(ggplot2)
sampledata<-data.frame(x=1:10,y=1:10)
graphique2<-ggplot(sampledata)+
  geom_point(aes(x,y))
\dontrun{
sortunsvg(graphique2,2,2,"graphiquedetest")
}
}
