% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{get_sources}
\alias{get_sources}
\title{Get index of available sources}
\usage{
get_sources(category = NULL, language = NULL, country = NULL)
}
\arguments{
\item{category}{(Optional) A string. Find sources that display news of a
specific category. Possible options are: "business", "entertainment",
"general", "health", "science", "sports", and "technology". By default, the
API returns sources of all categories.}

\item{language}{(Optional) A string. Find sources that display news in a specific language.
Possible options are: Arabic, German, English, Spanish, French, Hebrew, Italian
Dutch, Norwegian, Portuguese, Russian, Swedish, Urdu, and Chinese. The language
cam also be specified using ISO codes, e.g. "DE" for German or "ES" for Spanish.
By default, the API returns sources in all languages.}

\item{country}{(Optional) A string. The 2-letter ISO 3166-1 code of the country you
want to get headlines for, such as "US" or "FR". This parameter cannot be mixed with
the sources parameter.}
}
\value{
A data frame of news publishers, with their ID as it should be queried
in a NewsAPI request, a short description, their URL, their category, their
language, and their country.
}
\description{
Returns the subset of news publishers that top headlines are available from.
It's mainly a convenience function that can be used to keep track of the
publishers available on the API.
}
\examples{
get_sources(language = "German")

}
